fix a compilation warning (wrong type was used)
git-svn-id: http://root.cern.ch/svn/root/trunk@36768 27541ba8-7e3a-0410-8455-c3a389f83636

+++ math/mathcore/inc/Math/AdaptiveIntegratorMultiDim.h
@@ -103,13 +103,13 @@ public:
    void SetAbsTolerance(double absTol);
 
    ///set workspace size 
-   void SetSize(double size) { fSize = size; }
+   void SetSize(unsigned int size) { fSize = size; }
 
    ///set min points
-   void SetMinPts(double n) { fMinPts = n; }
+   void SetMinPts(unsigned int n) { fMinPts = n; }
 
    ///set max points
-   void SetMaxPts(double n) { fMaxPts = n; }
+   void SetMaxPts(unsigned int n) { fMaxPts = n; }
 
    /// set the options 
    void SetOptions(const ROOT::Math::IntegratorMultiDimOptions & opt);
